# Conflicts:
#	loader/LoaderAndLayerInterface.md
#	vktrace/src/vktrace_layer/vktrace_lib_helpers.h

+++ tests/blit_tests.cpp
@@ -2,23 +2,17 @@
 // Copyright (C) 2015-2016 Valve Corporation
 // Copyright (C) 2015-2016 LunarG, Inc.
 //
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the "Software"),
-// to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
-// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-// DEALINGS IN THE SOFTWARE.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 //
 // Author: Chia-I Wu <olv@lunarg.com>
 // Author: Jeremy Hayes <jeremy@lunarg.com>
@@ -290,7 +284,7 @@ std::vector<uint8_t> ImageChecker::pattern_hash(const VkImageSubresourceLayers &
     };
     unsigned long hash = 5381;
 
-    for (int32_t i = 0; i < ARRAY_SIZE(input); i++)
+    for (size_t i = 0; i < ARRAY_SIZE(input); i++)
         hash = ((hash << 5) + hash) + input[i];
 
     const uint8_t output[4] = { HASH_BYTES(hash) };
@@ -639,7 +633,7 @@ TEST_F(VkCmdFillBufferTest, MultiAlignments)
     VkMemoryPropertyFlags reqs = VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT;
 
     cmd_.begin();
-    for (int i = 0; i < ARRAY_SIZE(bufs); i++) {
+    for (size_t i = 0; i < ARRAY_SIZE(bufs); i++) {
         bufs[i].init_as_dst(dev_, size, reqs);
         vkCmdFillBuffer(cmd_.handle(), bufs[i].handle(), 0, size, 0x11111111);
         size <<= 1;
@@ -649,7 +643,7 @@ TEST_F(VkCmdFillBufferTest, MultiAlignments)
     submit_and_done();
 
     size = 4;
-    for (int i = 0; i < ARRAY_SIZE(bufs); i++) {
+    for (size_t i = 0; i < ARRAY_SIZE(bufs); i++) {
         const uint32_t *data = static_cast<const uint32_t *>(bufs[i].memory().map());
         VkDeviceSize offset;
         for (offset = 0; offset < size; offset += 4)
@@ -753,10 +747,10 @@ TEST_F(VkCmdCopyBufferTest, MultiAlignments)
     submit_and_done();
 
     data = static_cast<uint8_t *>(dst.memory().map());
-    for (int i = 0; i < ARRAY_SIZE(regions); i++) {
+    for (size_t i = 0; i < ARRAY_SIZE(regions); i++) {
         const VkBufferCopy &r = regions[i];
 
-        for (int j = 0; j < r.size; j++) {
+        for (uint64_t j = 0; j < r.size; j++) {
             EXPECT_EQ(r.srcOffset + j, data[r.dstOffset + j]) <<
                 "Region is: " << i << "\n" <<
                 "Offset is: " << r.dstOffset + j;
@@ -788,7 +782,7 @@ TEST_F(VkCmdCopyBufferTest, RAWHazard)
     err = vkResetEvent(dev_.handle(), event);
     ASSERT_VK_SUCCESS(err);
 
-    for (int i = 0; i < ARRAY_SIZE(bufs); i++) {
+    for (size_t i = 0; i < ARRAY_SIZE(bufs); i++) {
         bufs[i].init_as_src_and_dst(dev_, 4, reqs);
 
         uint32_t *data = static_cast<uint32_t *>(bufs[i].memory().map());
